refactor: move dependent resource code to own file
